Move test helpers to shared Cache scope
